Refactor ui package structure
