+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.WarOfSuits.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name="com.cianjansen.warofsuits.ui.activity.main.MainActivity"
             android:exported="true"
             android:configChanges="orientation"
             android:screenOrientation="portrait">
@@ -26,19 +26,19 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.game.GameActivity"
-            android:parentActivityName=".ui.main.MainActivity"
+            android:name="com.cianjansen.warofsuits.ui.activity.game.GameActivity"
+            android:parentActivityName="com.cianjansen.warofsuits.ui.activity.main.MainActivity"
             android:configChanges="orientation"
             android:screenOrientation="portrait"/>
         <activity
-            android:name=".ui.victory.VictoryActivity"
-            android:parentActivityName=".ui.main.MainActivity"
+            android:name="com.cianjansen.warofsuits.ui.activity.victory.VictoryActivity"
+            android:parentActivityName="com.cianjansen.warofsuits.ui.activity.main.MainActivity"
             android:configChanges="orientation"
             android:screenOrientation="portrait"/>
         <activity
             android:label="@string/summary_title"
-            android:name=".ui.summary.SummaryActivity"
-            android:parentActivityName=".ui.victory.VictoryActivity"
+            android:name="com.cianjansen.warofsuits.ui.activity.summary.SummaryActivity"
+            android:parentActivityName="com.cianjansen.warofsuits.ui.activity.victory.VictoryActivity"
             android:configChanges="orientation"
             android:screenOrientation="portrait"/>
     </application>
