+++ app/src/main/java/com/cianjansen/warofsuits/ui/activity/base/BasePresenter.kt
@@ -1,0 +1,3 @@
+package com.cianjansen.warofsuits.ui.activity.base
+
+interface BasePresenter
