+++ app/src/main/java/com/cianjansen/warofsuits/ui/activity/base/BaseView.kt
@@ -1,4 +1,4 @@
-package com.cianjansen.warofsuits.ui.base
+package com.cianjansen.warofsuits.ui.activity.base
 
 interface BaseView<T> {
     fun setPresenter(presenter : T)
