+++ app/src/main/java/com/cianjansen/warofsuits/ui/activity/game/GameContract.kt
@@ -1,8 +1,8 @@
-package com.cianjansen.warofsuits.ui.game
+package com.cianjansen.warofsuits.ui.activity.game
 
 import com.cianjansen.warofsuits.model.TurnSummary
-import com.cianjansen.warofsuits.ui.base.BasePresenter
-import com.cianjansen.warofsuits.ui.base.BaseView
+import com.cianjansen.warofsuits.ui.activity.base.BasePresenter
+import com.cianjansen.warofsuits.ui.activity.base.BaseView
 import com.cianjansen.warofsuits.model.PlayingCard
 
 class GameContract {
