+++ app/src/main/java/com/cianjansen/warofsuits/ui/activity/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.cianjansen.warofsuits.ui.main
+package com.cianjansen.warofsuits.ui.activity.main
 
 import android.os.Bundle
 import android.view.Menu
@@ -6,8 +6,8 @@ import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import com.cianjansen.warofsuits.R
 import com.cianjansen.warofsuits.databinding.ActivityMainBinding
-import com.cianjansen.warofsuits.ui.game.GameActivity
-import com.cianjansen.warofsuits.ui.victory.VictoryActivity
+import com.cianjansen.warofsuits.ui.activity.game.GameActivity
+import com.cianjansen.warofsuits.ui.activity.victory.VictoryActivity
 import com.google.android.material.snackbar.Snackbar
 
 /**
