+++ app/src/main/java/com/cianjansen/warofsuits/ui/activity/main/MainPresenter.kt
@@ -1,4 +1,4 @@
-package com.cianjansen.warofsuits.ui.main
+package com.cianjansen.warofsuits.ui.activity.main
 
 /**
  * The presenter for the main activity. Handles user input delegation
