--- app/src/main/java/com/cianjansen/warofsuits/ui/base/BasePresenter.kt
@@ -1,3 +1,0 @@
-package com.cianjansen.warofsuits.ui.base
-
-interface BasePresenter
