+++ app/src/main/res/layout/activity_summary.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MainActivity">
+    tools:context=".ui.activity.main.MainActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/ablAppBarLayout"
